Add module auto-formatter
Try:

python3 isle.fmt --help

(Fun fact: I just realised the process was source -> AST -> bytecode ->
AST -> source, while it could have been source -> AST -> source, so I
quickly changed that before writing this commit. As a side benefit,
augmented assignment (and possibly more syntactic sugar) is now
preserved.)

Comments are not preserved yet, so this is definitely not as production
ready as the rest of Isle, not that Isle itself is production ready,
but you get the point.